قبل النوم change VA status

+++ src/main/java/com/javainuse/springbootsecurity/repository/AssociationRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 
 public interface AssociationRepo extends JpaRepository<Association, Integer> {
-    Association findById(int id);
+    Association findById(Long id);
     Association findByName(String name);
 
 }
